refactor(theme-chalk): upgrade to ts & renive gulp-mincss with gulp-clean-css
