+++ package.json
@@ -19,7 +19,7 @@
     "build:utils": "cross-env BABEL_ENV=utils babel packages/utils --extensions .ts --out-dir lib/utils",
     "build:locale": "cross-env BABEL_ENV=utils babel packages/locale --extensions .ts --out-dir lib/locale",
     "build:locale-umd": "node ./build/build-locale.js",
-    "build:theme": "rimraf packages/theme-chalk/lib && gulp build --gulpfile packages/theme-chalk/gulpfile.js && cp-cli packages/theme-chalk/lib lib/theme-chalk && rimraf packages/theme-chalk/lib",
+    "build:theme": "rimraf packages/theme-chalk/lib && gulp build --gulpfile packages/theme-chalk/gulpfile.ts && cp-cli packages/theme-chalk/lib lib/theme-chalk && rimraf packages/theme-chalk/lib",
     "build:helper": "node build/build-helper.js",
     "build:indices": "node build/build-indices.js",
     "format": "yarn format:scss",
@@ -101,6 +101,7 @@
     "transliteration": "^2.1.11",
     "ts-loader": "^8.0.3",
     "ts-morph": "^11.0.3",
+    "ts-node": "^10.1.0",
     "typescript": "^4.0.2",
     "url-loader": "^4.1.0",
     "vue": "3.1.1",
